Lemon: Fixed compiler warning on Visual C++.
FossilOrigin-Name: 892c5f377a983fb440f421b08864780d4d4834e5

+++ tool/lemon.c
@@ -3249,7 +3249,7 @@ PRIVATE char *append_str(const char *zText, int n, int p1, int p2){
     }
     n = lemonStrlen(zText);
   }
-  if( n+sizeof(zInt)*2+used >= alloced ){
+  if( (int) (n+sizeof(zInt)*2+used) >= alloced ){
     alloced = n + sizeof(zInt)*2 + used + 200;
     z = (char *) realloc(z,  alloced);
   }
